refactor: move cot callback to internal
